requirements.txt
firebase_admin==6.1.0
Flask==1.1.2
Flask_Cors==3.0.10
Pyrebase4==4.6.0
Werkzeug==2.0.2
